+++ lib/domain/settings/row_model.dart
